+++ cppe/src/IceE/BasicStream.cpp
@@ -640,7 +640,7 @@ IceInternal::BasicStream::read(pair<const Byte*, const Byte*>& v)
 void
 IceInternal::BasicStream::write(const Byte* begin, const Byte* end)
 {
-    Int sz = end - begin;
+    Int sz = static_cast<Int>(end - begin);
     writeSize(sz);
     if(sz > 0)
     {
@@ -666,7 +666,7 @@ IceInternal::BasicStream::write(const vector<bool>& v)
 void
 IceInternal::BasicStream::write(const bool* begin, const bool* end)
 {
-    Int sz = end - begin;
+    Int sz = static_cast<Int>(end - begin);
     writeSize(sz);
     if(sz > 0)
     {
@@ -772,7 +772,7 @@ IceInternal::BasicStream::read(Short& v)
 void
 IceInternal::BasicStream::write(const Short* begin, const Short* end)
 {
-    Int sz = end - begin;
+    Int sz = static_cast<Int>(end - begin);
     writeSize(sz);
     if(sz > 0)
     {
@@ -871,7 +871,7 @@ IceInternal::BasicStream::read(Int& v)
 void
 IceInternal::BasicStream::write(const Int* begin, const Int* end)
 {
-    Int sz = end - begin;
+    Int sz = static_cast<Int>(end - begin);
     writeSize(sz);
     if(sz > 0)
     {
@@ -990,7 +990,7 @@ IceInternal::BasicStream::read(Long& v)
 void
 IceInternal::BasicStream::write(const Long* begin, const Long* end)
 {
-    Int sz = end - begin;
+    Int sz = static_cast<Int>(end - begin);
     writeSize(sz);
     if(sz > 0)
     {
@@ -1101,7 +1101,7 @@ IceInternal::BasicStream::read(Float& v)
 void
 IceInternal::BasicStream::write(const Float* begin, const Float* end)
 {
-    Int sz = end - begin;
+    Int sz = static_cast<Int>(end - begin);
     writeSize(sz);
     if(sz > 0)
     {
@@ -1220,7 +1220,7 @@ IceInternal::BasicStream::read(Double& v)
 void
 IceInternal::BasicStream::write(const Double* begin, const Double* end)
 {
-    Int sz = end - begin;
+    Int sz = static_cast<Int>(end - begin);
     writeSize(sz);
     if(sz > 0)
     {
@@ -1313,7 +1313,7 @@ IceInternal::BasicStream::write(const string& v)
 void
 IceInternal::BasicStream::write(const string* begin, const string* end)
 {
-    Int sz = end - begin;
+    Int sz = static_cast<Int>(end - begin);
     writeSize(sz);
     if(sz > 0)
     {
